+++ coralogix/resource_coralogix_team.go
@@ -4,12 +4,16 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"math"
 	"strconv"
 
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64planmodifier"
 	"terraform-provider-coralogix/coralogix/clientset"
 	teams "terraform-provider-coralogix/coralogix/clientset/grpc/teams"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/float64planmodifier"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 
 	"google.golang.org/protobuf/encoding/protojson"
 
@@ -24,6 +28,8 @@ import (
 
 var (
 	createTeamURL = "com.coralogixapis.aaa.organisations.v2.TeamService/CreateTeamInOrg"
+	updateTeamURL = "com.coralogixapis.aaa.organisations.v2.TeamService/UpdateTeam"
+	deleteTeamURL = "com.coralogixapis.aaa.organisations.v2.TeamService/DeleteTeam"
 )
 
 func NewTeamResource() resource.Resource {
@@ -75,19 +81,20 @@ func (r *TeamResource) Schema(ctx context.Context, req resource.SchemaRequest, r
 				Required:            true,
 				MarkdownDescription: "Team name.",
 			},
-			"team_admins_emails": schema.SetAttribute{
-				ElementType:         types.StringType,
-				Required:            true,
-				MarkdownDescription: "Team admins emails.",
-			},
 			"retention": schema.Int64Attribute{
-				Required:            true,
-				MarkdownDescription: "Team retention.",
-			},
-			"send_data_key": schema.StringAttribute{
 				Computed:            true,
-				Sensitive:           true,
-				MarkdownDescription: "Team send data key. Generated on creation.",
+				MarkdownDescription: "Team retention.",
+				PlanModifiers: []planmodifier.Int64{
+					int64planmodifier.UseStateForUnknown(),
+				},
+			},
+			"daily_quota": schema.Float64Attribute{
+				Computed:            true,
+				Optional:            true,
+				MarkdownDescription: "Team quota. Optional, Default daily quota is 0.01 units/day.",
+				PlanModifiers: []planmodifier.Float64{
+					float64planmodifier.UseStateForUnknown(),
+				},
 			},
 		},
 		MarkdownDescription: "Coralogix Team.",
@@ -95,11 +102,10 @@ func (r *TeamResource) Schema(ctx context.Context, req resource.SchemaRequest, r
 }
 
 type TeamResourceModel struct {
-	ID               types.String `tfsdk:"id"`
-	Name             types.String `tfsdk:"name"`
-	TeamAdminsEmails types.Set    `tfsdk:"team_admins_emails"` //types.String
-	Retention        types.Int64  `tfsdk:"retention"`
-	SendDataKey      types.String `tfsdk:"send_data_key"`
+	ID         types.String  `tfsdk:"id"`
+	Name       types.String  `tfsdk:"name"`
+	Retention  types.Int64   `tfsdk:"retention"`
+	DailyQuota types.Float64 `tfsdk:"daily_quota"`
 }
 
 func (r *TeamResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
@@ -110,7 +116,7 @@ func (r *TeamResource) Create(ctx context.Context, req resource.CreateRequest, r
 		return
 	}
 
-	createTeamReq, diags := extractCreateTeam(ctx, plan)
+	createTeamReq, diags := extractCreateTeam(plan)
 	if diags.HasError() {
 		resp.Diagnostics = diags
 		return
@@ -134,33 +140,46 @@ func (r *TeamResource) Create(ctx context.Context, req resource.CreateRequest, r
 		return
 	}
 	log.Printf("[INFO] Submitted new team: %s", protojson.Format(createTeamResp.GetTeamId()))
-	plan = flattenTeam(createTeamReq, createTeamResp)
+
+	getTeamReq := &teams.GetTeamRequest{
+		TeamId: createTeamResp.GetTeamId(),
+	}
+	getTeamResp, err := r.client.GetTeam(ctx, getTeamReq)
+	if err != nil {
+		log.Printf("[ERROR] Received error: %s", err.Error())
+		resp.Diagnostics.AddError(
+			"Error reading Team",
+			formatRpcErrors(err, getTeamURL, protojson.Format(getTeamReq)),
+		)
+		return
+	}
+	log.Printf("[INFO] Received Team: %s", protojson.Format(getTeamResp))
+	state := flattenTeam(getTeamResp)
 
 	// Set state to fully populated data
-	diags = resp.State.Set(ctx, plan)
+	diags = resp.State.Set(ctx, state)
 	resp.Diagnostics.Append(diags...)
 }
 
-func extractCreateTeam(ctx context.Context, plan *TeamResourceModel) (*teams.CreateTeamInOrgRequest, diag.Diagnostics) {
-	emails, diags := typeStringSliceToStringSlice(ctx, plan.TeamAdminsEmails.Elements())
-	if diags.HasError() {
-		return nil, diags
+func extractCreateTeam(plan *TeamResourceModel) (*teams.CreateTeamInOrgRequest, diag.Diagnostics) {
+	var dailyQuota *float64
+	if !(plan.DailyQuota.IsUnknown() || plan.DailyQuota.IsNull()) {
+		dailyQuota = new(float64)
+		*dailyQuota = plan.DailyQuota.ValueFloat64()
 	}
 
 	return &teams.CreateTeamInOrgRequest{
-		TeamName:        plan.Name.ValueString(),
-		Retention:       int32(plan.Retention.ValueInt64()),
-		TeamAdminsEmail: emails,
+		TeamName:   plan.Name.ValueString(),
+		DailyQuota: dailyQuota,
 	}, nil
 }
 
-func flattenTeam(req *teams.CreateTeamInOrgRequest, resp *teams.CreateTeamInOrgResponse) *TeamResourceModel {
+func flattenTeam(resp *teams.GetTeamResponse) *TeamResourceModel {
 	return &TeamResourceModel{
-		ID:               types.StringValue(strconv.Itoa(int(resp.GetTeamId().GetId()))),
-		Name:             types.StringValue(req.GetTeamName()),
-		TeamAdminsEmails: stringSliceToTypeStringSet(req.GetTeamAdminsEmail()),
-		Retention:        types.Int64Value(int64(req.GetRetention())),
-		SendDataKey:      types.StringValue(resp.GetSendDataKey()),
+		ID:         types.StringValue(strconv.Itoa(int(resp.GetTeamId().GetId()))),
+		Name:       types.StringValue(resp.GetTeamName()),
+		Retention:  types.Int64Value(int64(resp.GetRetention())),
+		DailyQuota: types.Float64Value(math.Round(resp.GetDailyQuota()*1000) / 1000),
 	}
 }
 
@@ -172,49 +191,155 @@ func (r *TeamResource) Read(ctx context.Context, req resource.ReadRequest, resp 
 		return
 	}
 
+	intId, err := strconv.Atoi(plan.ID.ValueString())
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error parsing Team ID",
+			fmt.Sprintf("Error parsing Team ID: %s", err.Error()),
+		)
+		return
+	}
+	getTeamReq := &teams.GetTeamRequest{
+		TeamId: &teams.TeamId{
+			Id: uint32(intId),
+		},
+	}
+	log.Printf("[INFO] Reading Team: %s", protojson.Format(getTeamReq))
+	getTeamResp, err := r.client.GetTeam(ctx, getTeamReq)
+	if err != nil {
+		log.Printf("[ERROR] Received error: %s", err.Error())
+		if status.Code(err) == codes.NotFound {
+			plan.ID = types.StringNull()
+			resp.Diagnostics.AddWarning(
+				fmt.Sprintf("Team %q is in state, but no longer exists in Coralogix backend", intId),
+				fmt.Sprintf("%q will be recreated when you apply", intId),
+			)
+		} else {
+			resp.Diagnostics.AddError(
+				"Error reading Team",
+				formatRpcErrors(err, getTeamURL, protojson.Format(getTeamReq)),
+			)
+		}
+		return
+	}
+	log.Printf("[INFO] Received Team: %s", protojson.Format(getTeamResp))
+
+	state := flattenTeam(getTeamResp)
 	// Set state to fully populated data
-	diags = resp.State.Set(ctx, plan)
+	diags = resp.State.Set(ctx, state)
 	resp.Diagnostics.Append(diags...)
 }
 
 func (r *TeamResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var plan, state *TeamResourceModel
-	diags := req.State.Get(ctx, &plan)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-	diags = req.Plan.Get(ctx, &state)
+	var plan *TeamResourceModel
+	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	if plan.Name.ValueString() != state.Name.ValueString() {
-		resp.Diagnostics.AddError(
-			"Error updating Team",
-			"Team name cannot be updated.",
-		)
+	updateReq, diags := extractUpdateTeam(plan)
+	if diags.HasError() {
+		resp.Diagnostics = diags
+		return
+	}
+	log.Printf("[INFO] Updating Team: %s", protojson.Format(updateReq))
+
+	_, err := r.client.UpdateTeam(ctx, updateReq)
+	if err != nil {
+		log.Printf("[ERROR] Received error: %s", err.Error())
+		if status.Code(err) == codes.PermissionDenied || status.Code(err) == codes.Unauthenticated {
+			resp.Diagnostics.AddError(
+				"Error updating Team",
+				fmt.Sprintf("permission denied for url - %s\ncheck your org-key and permissions", updateTeamURL),
+			)
+		} else {
+			resp.Diagnostics.AddError(
+				"Error updating Team",
+				formatRpcErrors(err, updateTeamURL, protojson.Format(updateReq)),
+			)
+		}
+
+		return
 	}
 
-	if plan.Retention.ValueInt64() != state.Retention.ValueInt64() {
-		resp.Diagnostics.AddError(
-			"Error updating Team",
-			"Team retention cannot be updated.",
-		)
-	}
+	log.Printf("[INFO] Updated team: %s", plan.ID.ValueString())
 
-	if !plan.TeamAdminsEmails.Equal(state.TeamAdminsEmails) {
-		resp.Diagnostics.AddError(
-			"Error updating Team",
-			"Team admins cannot be updated.",
-		)
+	getTeamReq := &teams.GetTeamRequest{
+		TeamId: updateReq.GetTeamId(),
 	}
+	getTeamResp, err := r.client.GetTeam(ctx, getTeamReq)
+	if err != nil {
+		log.Printf("[ERROR] Received error: %s", err.Error())
+		resp.Diagnostics.AddError(
+			"Error reading Team",
+			formatRpcErrors(err, getTeamURL, protojson.Format(getTeamReq)),
+		)
+		return
+	}
+	log.Printf("[INFO] Received Team: %s", protojson.Format(getTeamResp))
+	state := flattenTeam(getTeamResp)
+
+	// Set state to fully populated data
+	diags = resp.State.Set(ctx, state)
+	resp.Diagnostics.Append(diags...)
+}
+
+func extractUpdateTeam(plan *TeamResourceModel) (*teams.UpdateTeamRequest, diag.Diagnostics) {
+	dailyQuota := new(float64)
+	*dailyQuota = plan.DailyQuota.ValueFloat64()
+
+	id, err := strconv.Atoi(plan.ID.ValueString())
+	if err != nil {
+		return nil, diag.Diagnostics{diag.NewErrorDiagnostic("Error converting team id to int", err.Error())}
+	}
+	teamId := &teams.TeamId{Id: uint32(id)}
+
+	teamName := new(string)
+	*teamName = plan.Name.ValueString()
+
+	return &teams.UpdateTeamRequest{
+		TeamId:     teamId,
+		TeamName:   teamName,
+		DailyQuota: dailyQuota,
+	}, nil
 }
 
 func (r *TeamResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	resp.Diagnostics.AddError(
-		"Delete not supported",
-		"Delete is not supported for this resource.",
-	)
+	var state *TeamResourceModel
+	diags := req.State.Get(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	log.Printf("[INFO] Deleting Team: %s", state.ID.ValueString())
+	id, err := strconv.Atoi(state.ID.ValueString())
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error deleting Team",
+			fmt.Sprintf("Error converting team id to int: %s", err.Error()),
+		)
+		return
+	}
+
+	deleteReq := &teams.DeleteTeamRequest{TeamId: &teams.TeamId{Id: uint32(id)}}
+	log.Printf("[INFO] Deleting Team: %s", protojson.Format(deleteReq))
+	_, err = r.client.DeleteTeam(ctx, deleteReq)
+	if err != nil {
+		log.Printf("[ERROR] Received error: %s", err.Error())
+		if status.Code(err) == codes.PermissionDenied || status.Code(err) == codes.Unauthenticated {
+			resp.Diagnostics.AddError(
+				"Error deleting Team",
+				fmt.Sprintf("permission denied for url - %s\ncheck your org-key and permissions", deleteTeamURL),
+			)
+		} else {
+			resp.Diagnostics.AddError(
+				"Error deleting Team",
+				formatRpcErrors(err, deleteTeamURL, protojson.Format(deleteReq)),
+			)
+		}
+		return
+	}
+	log.Printf("[INFO] Deleted team: %s", state.ID.ValueString())
 }