+++ CHANGELOG.md
@@ -488,3 +488,8 @@ Bug fixing:
 New Features:
 #### resource/coralogix_dashboards_folder
 * Adding support for `parent_id`.
+
+## Release 1.11.13
+Breaking Changes:
+#### resource/coralogix_team and resource/coralogix_moving_quota
+* `coralogix_moving_quota` was removed, and the `coralogix_team` resource was changed to support setting of daily-quota.